refactor: kanban eventhander 분리
