+++ index.js
@@ -1,4 +1,5 @@
 import App from './app.js';
 import dataController from './src/js/services/index.js';
+import { $ } from './src/js/utils/querySelector.js';
 
-new App(document.querySelector('.team-view'), dataController);
+new App($('.team-view'), dataController);
