+++ src/js/kanban/components/teamHeader.js
